Move keyvault/7.1-preview out of preview since it should be stable
