+++ CHANGELOG.md
@@ -41,6 +41,8 @@
 
 - package `github.com/Azure/azure-sdk-for-go/services/preview/hdinsight/mgmt/2018-06-01-preview/hdinsight` has been renamed to `github.com/Azure/azure-sdk-for-go/services/hdinsight/mgmt/2018-06-01/hdinsight`
 
+- package `github.com/Azure/azure-sdk-for-go/services/preview/keyvault/v7.1-preview/keyvault` has been renamed to `github.com/Azure/azure-sdk-for-go/services/keyvault/v7.1/keyvault`
+
 Generated from https://github.com/Azure/azure-rest-api-specs/tree/93106ff722d481c30a9a4ae089564635c10e9401
 
 ## `v46.4.0`
